teams_agent/requirements.txt
google-cloud-aiplatform[adk,agent_engines]
google-adk>=1.12.0
google-generativeai>=0.8.5
python-dotenv>=1.1.1
absl-py>=2.3.1
aiohttp>=3.12.15
requests>=2.31.0
cloudpickle
pydantic
